Splitting rules by property and method to allow for easier customisation later. Moved reused code into a class of it's own
