+++ tests/rules/method.js
@@ -1,0 +1,111 @@
+/* global require */
+/**
+ * @fileoverview Test for no-unsanitized rule
+ * @author Frederik Braun et al.
+ * @copyright 2015-2017 Mozilla Corporation. All rights reserved
+ */
+
+//------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+const rule = require("../../lib/rules/method");
+const RuleTester = require("eslint").RuleTester;
+
+//------------------------------------------------------------------------------
+// Tests
+//------------------------------------------------------------------------------
+
+const eslintTester = new RuleTester();
+
+eslintTester.run("method", rule, {
+
+    // Examples of code that should not trigger the rule
+    // XXX this does not find z['innerHTML'] and the like.
+
+    valid: [
+        // testing unwrapSafeHTML spread
+        {
+            code: "this.imeList.innerHTML = Sanitizer.unwrapSafeHTML(...listHtml);",
+            parserOptions: { ecmaVersion: 6 }
+        },
+
+        // tests for insertAdjacentHTML calls
+        {
+            code: "n.insertAdjacentHTML('afterend', 'meh');",
+            parserOptions: { ecmaVersion: 6 }
+        },
+        {
+            code: "n.insertAdjacentHTML('afterend', `<br>`);",
+            parserOptions: { ecmaVersion: 6 }
+        },
+        {
+            code: "n.insertAdjacentHTML('afterend', Sanitizer.escapeHTML`${title}`);",
+            parserOptions: { ecmaVersion: 6 }
+        },
+
+        // document.write/writeln
+        {
+            code: "document.write('lulz');",
+            parserOptions: { ecmaVersion: 6 }
+        },
+        {
+            code: "document.write();",
+            parserOptions: { ecmaVersion: 6 }
+        },
+        {
+            code: "document.writeln(Sanitizer.escapeHTML`<em>${evil}</em>`);",
+            parserOptions: { ecmaVersion: 6 }
+        },
+    ],
+
+    // Examples of code that should trigger the rule
+    invalid: [
+        /* XXX Do NOT change the error strings below without review from freddy:
+         * The strings are optimized for SEO and understandability.
+         * The developer can search for them and will find this MDN article:
+         *  https://developer.mozilla.org/en-US/Firefox_OS/Security/Security_Automation
+         */
+
+        // insertAdjacentHTML examples
+        {
+            code: "node.insertAdjacentHTML('beforebegin', htmlString);",
+            errors: [
+                {
+                    message: "Unsafe call to insertAdjacentHTML",
+                    type: "CallExpression"
+                }
+            ]
+        },
+        {
+            code: "node.insertAdjacentHTML('beforebegin', template.getHTML());",
+            errors: [
+                {
+                    message: "Unsafe call to insertAdjacentHTML",
+                    type: "CallExpression"
+                }
+            ]
+        },
+
+        // document.write / writeln
+        {
+            code: "document.write('<span>'+ htmlInput + '</span>');",
+            errors: [
+                {
+                    message: "Unsafe call to document.write",
+                    type: "CallExpression"
+                }
+            ]
+        },
+        {
+            code: "document.writeln(evil);",
+            errors: [
+                {
+                    message: "Unsafe call to document.writeln",
+                    type: "CallExpression"
+                }
+            ]
+        },
+
+    ]
+});
